增加： 1、增加软键盘：全键盘支持拼音中文输入 小键盘支持数字输入
修改：
1、配合软键盘，修改输入控件的类型

+++ src/dataCommit.h
@@ -225,7 +225,7 @@ extern short drughttype;
 extern short gain_autopilot;
 extern float max_rudangle;
 extern int gainlevel;
-extern int set_maxRudderAngle;
+extern float set_maxRudderAngle;
 extern bool flag_weathercomp;
 extern unsigned int avlb_power;
 extern unsigned int real_power;
